requirements.txt
biopython
click
numpy
pandas
pydantic
scikit-image
